Have the configuration change BaseModel's base content directory
Given that the Config class should be driving everything, it makes sense
to me to have the Config class follow through with making some
configuration changes.

I could also very easily see this as being a Not Wanted Thing™, because
it means a data class is doing more than just holding data.