Move model blueprints to a Controller class
This is a big chunk of work, and it actually adds a significant number
of minor changes too, because I'm awful at splitting things up into
smaller commits when I'm flowing. grr.

So the big one: model blueprints are decoupled from the BaseModel class.
This was mostly done because it seems like a significantly better way to
go about having blueprints (the reason I coupled them together at first
was mostly laziness).

This itself has two new, smaller features baked in:

- Editing a page/game's filename
- A index page showing all pages/games

A third, medium sized feature is that creating new models is now done in
the same form as editing models, instead of having a two-step create
then edit process.

Additional niceties that aren't full features in my mind:

- The edit form now just uses the regular browser/form API instead of
  sending JSON data

- The JS around async edit form submissions was cleaned up, because it
  was missed in the last pass through

- The global sidebar now just has a single link for pages/games, instead
  of listing all pages and games

Coupled with these changes is the addition of a significantly more
robust testing suite than the previous, basically non-existent one.
Someone wanting to make changes to the application should now be able to
determine if they broke something in _most_ places just by running the
tests, instead of having to go through an entire flow.

(that said, the entire deployment procedure is non-tested, though I hope
to tackle that next, or after an image library frontend)