Add GeneratorService
GeneratorService is a replacement to the current "static" module, which
is used to build the actual site from the source files.

GeneratorService is an improvement because it now includes a rollback
feature, and now keeps five recent deploys around instead of only
creating one directory. It also swaps the new deploy in using a symlink
over just writing the contents to the "live" directory.