Don't rely on os.listdir to return sorted
Having only run the test suite on OS X before, there was never a reason
for me to doubt that os.listdir always returned sorted (which it does on
OS X). This is a bad assumption, and the tests failed when run on an
ext4 filesystem.

GeneratorService.deploys() should now return a sorted list of deploys,
instead of relying on os.listdir to do that