Completely rework Config to make it a global configuration object
The goal of this side endeavor was to (and in my belief, based on ensuring
the tests continue to pass, successfully has been) turning the server
into something that could easily be run with WSGI.

The tangent off of that goal required changing Config into a global
object, both for tests to function well, and for the application to be
dynamically configured in the same way it's currently configured (with a
single yaml file).

Unfortunately, this means a lot of idempotency has been lost, because of
basically everything relying on the global Config object, which can
change at basically any point in operation (but _shouldn't_, it only
does because of testing).

That said, I've made the executive decision that it's okay, because this
really isn't meant to be something other people could take components of
to build out their own fun stuff. Especially with the plans to include a
database (which is another driver for the app being easily WSGI runnable
(the end state makes the database easier to manage)).

Anyways, enjoy all these huge changes and a large paradigm shift.
Thankfully I'm the only developer right now, so nobody else really needs
to deal with this. Yet.

Other niceties added:

-   add procfile, for foreman-y goodness
-   link readme to travis