Create split between file-backed and database-backed controllers
This is probably just going to be a temporary measure, and long-term all
the models will be converted into database models. But I wanted to be
able to have the application continue to function for both
simultaneously.

If you ever deploy a commit between these two versions, you're doing
something very very very bad. But you definitely could, if you hate
yourself and want to have some models backed in a database and some
backed by files.

But don't do that.